refactor: 시간 포맷 중 초단위는 제거
