+++ web/src/components/dashboard/SubmittedExamContent.tsx
@@ -1,0 +1,5 @@
+const SubmittedExamContent = () => {
+  return <div>SubmittedExamContent</div>;
+};
+
+export default SubmittedExamContent;
