+++ web/src/pages/dashboard/DashboardPage.tsx
@@ -1,12 +1,11 @@
 import AsyncBoundary from '@/components/AsyncBoundary';
-import useGetMyExamSummaries from '@/hooks/api/exam/useGetMyExamSummaries';
-import DraftExamCard from '@/components/dashboard/DraftExamCard';
-import PublishedExamCard from '@/components/dashboard/PublishedExamCard';
-import { Pagination, Tab, Tabs } from '@nextui-org/react';
+import { Tab, Tabs } from '@nextui-org/react';
 import { BsFillFileEarmarkCheckFill, BsFillSendCheckFill } from 'react-icons/bs';
 import { PiPencilLineBold } from 'react-icons/pi';
 import NewExamButton from '@/components/exams/NewExamButton';
-import { useState } from 'react';
+import DraftExamContent from '@/components/dashboard/DraftExamContent';
+import PublishedExamContent from '@/components/dashboard/PublishedExamContent';
+import SubmittedExamContent from '@/components/dashboard/SubmittedExamContent';
 
 const DashboardPage = () => {
   return (
@@ -69,80 +68,12 @@ const DashboardPage = () => {
           >
             <AsyncBoundary>
               <div className="text-2xl font-semibold mb-5">제출한 시험들</div>
-              <PublishedExamContent />
+              <SubmittedExamContent />
             </AsyncBoundary>
           </Tab>
         </Tabs>
       </div>
     </div>
-  );
-};
-
-const DraftExamContent = () => {
-  const [page, setPage] = useState(1);
-  const { data } = useGetMyExamSummaries('DRAFT', page - 1);
-  const { content, pageInfo } = data;
-
-  if (!content || pageInfo.totalElements === 0) {
-    return (
-      <div className="w-full text-center text-xl text-gray-500 mt-24">
-        출제 준비 중인 시험이 없습니다.
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-4 gap-5">
-        {content.map((exam) => {
-          return <DraftExamCard key={exam.id} exam={exam} />;
-        })}
-      </div>
-      <div className="flex justify-center mt-10">
-        <Pagination
-          color="primary"
-          showControls
-          showShadow
-          page={page}
-          total={pageInfo.totalPages}
-          onChange={(page) => setPage(page)}
-        />
-      </div>
-    </>
-  );
-};
-
-const PublishedExamContent = () => {
-  const [page, setPage] = useState(1);
-  const { data } = useGetMyExamSummaries('PUBLISHED', page - 1);
-  const { content, pageInfo } = data;
-
-  if (!content || pageInfo.totalElements === 0) {
-    return (
-      <div className="w-full text-center text-xl text-gray-500 mt-24">
-        출제 완료된 시험이 없습니다.
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-4 gap-5">
-        {content.map((exam) => {
-          return <PublishedExamCard key={exam.id} exam={exam} />;
-        })}
-      </div>
-      <div className="flex justify-center mt-10">
-        <Pagination
-          color="primary"
-          showControls
-          showShadow
-          page={page}
-          total={pageInfo.totalPages}
-          onChange={(page) => setPage(page)}
-        />
-      </div>
-    </>
   );
 };
